optimize parquet writing for postgres based dbs
