+++ vectordb_bench/backend/clients/api.py
@@ -1,5 +1,7 @@
 from abc import ABC, abstractmethod
 import pathlib
+import time
+import logging
 from enum import Enum
 from typing import Any, Type
 from contextlib import contextmanager
@@ -7,6 +9,8 @@ from ... import config
 
 from pydantic import BaseModel, validator, SecretStr
 
+
+log = logging.getLogger(__name__) 
 
 class MetricType(str, Enum):
     L2 = "L2"
@@ -215,3 +219,86 @@ class VectorDB(ABC):
         heavy load cases.
         """
         raise NotImplementedError
+
+class PgDB(VectorDB):
+    def load_parquets(self, parquet_files: list[str]) -> int:
+        import pyarrow.dataset as ds
+        from pgpq import ArrowToPostgresBinaryEncoder
+
+        total_start = time.time()
+        # load an arrow dataset
+        # arrow can load datasets from partitioned parquet files locally or in S3/GCS
+        # it handles buffering, matching globs, etc.
+        log.info(f"loading {len(parquet_files)} files")
+        log.info(f"loading files: {parquet_files}")
+        dataset = ds.dataset(parquet_files)
+
+        # create an encoder object which will do the encoding
+        # and give us the expected Postgres table schema
+        encoder = ArrowToPostgresBinaryEncoder(dataset.schema)
+        # get the expected Postgres destination schema
+        # note that this is _not_ the same as the incoming arrow schema
+        # and not necessarily the schema of your permanent table
+        # instead it's the schema of the data that will be sent over the wire
+        # which for example does not have timezones on any timestamps
+
+        # here, since embedding tables just have id and float4[] columns, 
+        # it is fine to actualy use this table as the final table
+        pg_schema = encoder.schema()
+        # assemble ddl for a temporary table
+        # it's often a good idea to bulk load into a temp table to:
+        # (1) Avoid indexes
+        # (2) Stay in-memory as long as possible
+        # (3) Be more flexible with types
+        #     (you can't load a SMALLINT into a BIGINT column without casting)
+
+        tmp_table_name = "_tmp_parquet_data"
+        typed_cols = [f'"{col_name}" {col.data_type.ddl()}' for col_name, col in pg_schema.columns]
+        cols = [col_name for col_name, _ in pg_schema.columns]
+        cols_joined = ','.join(cols)
+        typed_cols_joined = ','.join(typed_cols)
+        self.pg_session.execute(f'DROP TABLE IF EXISTS "{tmp_table_name}"')
+        ddl = f"CREATE TEMPORARY TABLE {tmp_table_name} ({typed_cols_joined})"
+        self.pg_session.execute(ddl)
+
+        log.debug(f"pg schema {pg_schema}")
+        log.debug(f"Assuming underlying postgres table was created with columns: {typed_cols}")
+
+        count = 0
+        i = 0
+        for batch in dataset.to_batches():
+            encoder = ArrowToPostgresBinaryEncoder(dataset.schema)
+            log.info(f"batch: {i} batch len: {len(batch)}")
+            # Write to buffer
+            write_start = time.time()
+            self.binary_f.truncate(0)
+            self.binary_f.seek(0)
+            self.binary_f.write(encoder.write_header())
+            self.binary_f.write(encoder.write_batch(batch))
+            self.binary_f.write(encoder.finish())
+            log.debug(f"Writing batch to buffer took {int(time.time() - write_start)}s")
+
+            # Copy to tmp table
+            copy_start = time.time()
+            self.binary_f.seek(0)
+            self.pg_session.copy_expert(f'COPY "{tmp_table_name}" ({cols_joined}) FROM STDIN WITH (FORMAT BINARY)', self.binary_f)
+            log.debug(f"Writing batch to postgres took {int(time.time() - copy_start)}s, total batch processing took: {int(time.time() - write_start)}s")
+            count += len(batch)
+            i += 1
+
+        # clean buffer
+        self.binary_f.seek(0)
+        self.binary_f.truncate(0)
+
+        # Copy to actual table
+        copy_start = time.time()
+        self.pg_session.execute(f'INSERT INTO "{self.table_name}" SELECT * FROM "{tmp_table_name}"')
+        self.pg_session.execute(f'ALTER TABLE "{self.table_name}" ADD CONSTRAINT lantern_pk PRIMARY KEY ("{self._primary_field}");')
+        self.pg_session.execute(f'ALTER TABLE "{self.table_name}" SET LOGGED')
+        self.pg_session.execute(f'DROP TABLE "{tmp_table_name}"')
+        log.debug(f"Copy data to main table took {int(time.time() - copy_start)}s, total insert took {int(time.time() - total_start)}s")
+        vacuum_start = time.time()
+        self.pg_session.execute(f'VACUUM FULL "{self.table_name}"')
+        log.debug(f"Table vacuum took {int(time.time() - vacuum_start)}s")
+
+        return count
