+++ vectordb_bench/backend/clients/lantern/lantern.py
@@ -8,12 +8,12 @@ from contextlib import contextmanager
 from typing import Any, Tuple, Type
 import psycopg2
 
-from ..api import MetricType, VectorDB, DBConfig, DBCaseConfig, IndexType
+from ..api import MetricType, PgDB, DBConfig, DBCaseConfig, IndexType
 from .config import LanternConfig, LanternIndexConfig
 
 log = logging.getLogger(__name__) 
 
-class Lantern(VectorDB):
+class Lantern(PgDB):
     def __init__(
         self,
         dim: int,
@@ -80,7 +80,6 @@ class Lantern(VectorDB):
         pass
 
     def optimize(self):
-        self.pg_session.execute(f'ALTER TABLE "{self.table_name}" ADD CONSTRAINT lantern_pk PRIMARY KEY ("{self._primary_field}");')
         index_param = self.case_config.index_param()
         # create vec index
         self._create_index(self.pg_session)
@@ -107,75 +106,10 @@ class Lantern(VectorDB):
     def _create_table(self, pg_session, dim):
         try:
             # create table
-            pg_session.execute(f'CREATE TABLE "{self.table_name}" ("{self._primary_field}" INT8, "{self._vector_field}" REAL[{dim}]);')
+            pg_session.execute(f'CREATE UNLOGGED TABLE "{self.table_name}" ("{self._primary_field}" INT8, "{self._vector_field}" REAL[{dim}]);')
         except Exception as e:
             log.warning(f"Failed to create lantern table: {self.table_name} error: {e}")
             raise e from None
-
-    def load_parquets(self, parquet_files: list[str]) -> int:
-        import pyarrow.dataset as ds
-        from pgpq import ArrowToPostgresBinaryEncoder
-
-        # load an arrow dataset
-        # arrow can load datasets from partitioned parquet files locally or in S3/GCS
-        # it handles buffering, matching globs, etc.
-        log.info(f"loading {len(parquet_files)} files")
-        log.info(f"loading files: {parquet_files}")
-        dataset = ds.dataset(parquet_files)
-
-        # create an encoder object which will do the encoding
-        # and give us the expected Postgres table schema
-        encoder = ArrowToPostgresBinaryEncoder(dataset.schema)
-        # get the expected Postgres destination schema
-        # note that this is _not_ the same as the incoming arrow schema
-        # and not necessarily the schema of your permanent table
-        # instead it's the schema of the data that will be sent over the wire
-        # which for example does not have timezones on any timestamps
-
-        # here, since embedding tables just have id and float4[] columns, 
-        # it is fine to actualy use this table as the final table
-        pg_schema = encoder.schema()
-        # assemble ddl for a temporary table
-        # it's often a good idea to bulk load into a temp table to:
-        # (1) Avoid indexes
-        # (2) Stay in-memory as long as possible
-        # (3) Be more flexible with types
-        #     (you can't load a SMALLINT into a BIGINT column without casting)
-
-        tmp_table_name = "_tmp_parquet_data"
-        typed_cols = [f'"{col_name}" {col.data_type.ddl()}' for col_name, col in pg_schema.columns]
-        cols = [col_name for col_name, _ in pg_schema.columns]
-        cols_joined = ','.join(cols)
-        typed_cols_joined = ','.join(typed_cols)
-
-        ddl = f"CREATE UNLOGGED TABLE {tmp_table_name} ({typed_cols_joined})"
-
-        self.pg_session.execute(f"DROP TABLE IF EXISTS {tmp_table_name}")
-        self.pg_session.execute(ddl)
-        log.debug(f"pg schema {pg_schema}")
-        log.debug(f"Assuming underlying postgres table was created with columns: {typed_cols} via a statement equivalent (or columnwise-type-castable) to'{ddl}'")
-
-        self.binary_f.truncate(0)
-        self.binary_f.seek(0)
-        copy = self.binary_f
-        copy.write(encoder.write_header())
-        batches = dataset.to_batches()
-        count = 0
-        for i, batch in enumerate(batches):
-            log.info(f"batch: {i} batch len: {len(batch)}")
-            b = encoder.write_batch(batch)
-            copy.write(b)
-            count += len(batch)
-
-        copy.write(encoder.finish())
-        self.binary_f.seek(0)
-        # todo:: the below can actually run in a separate thread/process, parallel to the above
-        log.info(f"Copying dataset into postgres...")
-        self.pg_session.copy_expert(f'COPY "{tmp_table_name}" ({cols_joined}) FROM STDIN WITH (FORMAT BINARY)', self.binary_f)
-        self.pg_session.execute(f'INSERT INTO "{self.table_name}" SELECT * FROM "{tmp_table_name}"')
-        self.pg_session.execute(f'DROP TABLE "{tmp_table_name}"')
-        self.pg_session.execute(f'VACUUM FULL "{self.table_name}"')
-        return count
 
     def insert_embeddings(
         self,
